# Exercise7/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Module5 Exercise7 Calculator",
    "description": "My Calculator Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Calculator",
      "description": "API for Module5 Exercise7 Calculus"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],

  "paths": {
    "/calculator/add": {
      "get": {
        "tags": ["Addition Service"],
        "summary": "Add 2 numbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "The First Number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "The Second number"
          }
        ],
        "responses": {
          "200": {
            "description": "This service allows you to add two numbers together"
          }
        }
      }
    },
    "/calculator/subtract": {
        "get": {
          "tags": ["Subtraction Service"],
          "summary": "Subtract Second Number from First Number",
          "parameters": [
            {
              "name": "num1",
              "in": "query",
              "description": "The First Number"
            },
            {
              "name": "num2",
              "in": "query",
              "description": "The Second number"
            }
          ],
          "responses": {
            "200": {
              "description": "This service allows you to subtract second number from first number"
            }
          }
        }
      },
      "/calculator/multiply": {
        "get": {
          "tags": ["Multiplication Service"],
          "summary": "Multiply 2 numbers",
          "parameters": [
            {
              "name": "num1",
              "in": "query",
              "description": "The First Number"
            },
            {
              "name": "num2",
              "in": "query",
              "description": "The Second number"
            }
          ],
          "responses": {
            "200": {
              "description": "This service allows you to multiply two numbers together"
            }
          }
        }
      },
      "/calculator/divide": {
        "get": {
          "tags": ["Division Service"],
          "summary": "Divide first number by second number",
          "parameters": [
            {
              "name": "num1",
              "in": "query",
              "description": "The First Number"
            },
            {
              "name": "num2",
              "in": "query",
              "description": "The Second number"
            }
          ],
          "responses": {
            "200": {
              "description": "This service allows you to divide first number by second number"
            }
          }
        }
      }
    
    
    
  }
}